Revert: LinuxKPI: switch jiffies and timer->expire to unsigned long
There are possible problems when jiffies (ticks) which still are int
wrap around.  Also given this did not touch every single place some
checks may be broken now.

Reported by:	markj
This reverts commit 438a37a3f81f8cd24b5d4df5e1bc777c94704b91.

+++ sys/compat/linuxkpi/common/src/linux_compat.c
@@ -1938,7 +1938,7 @@ linux_timer_callback_wrapper(void *context)
 }
 
 int
-mod_timer(struct timer_list *timer, unsigned long expires)
+mod_timer(struct timer_list *timer, int expires)
 {
 	int ret;
